% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDSVfilter.R
\name{MDSVfilter}
\alias{MDSVfilter}
\title{MDSV Filtering}
\usage{
MDSVfilter(N, K, data, para, ModelType = 0, LEVIER = FALSE)
}
\arguments{
\item{N}{An integer designing the number of components for the MDSV process}

\item{K}{An integer designing the number of states of each MDSV process component}

\item{data}{A univariate or bivariate data matrix. Can only be a matrix of 1 or 2 columns. If data has 2 columns, the first one has to be the log-returns and the second the realized variances.}

\item{para}{A vector of parameters use for the MDSV filtering on `data`. For more informations see Details.}

\item{ModelType}{An integer designing the type of model to be fit. \eqn{0} for univariate log-returns, \eqn{1} for univariate realized variances and \eqn{2} for joint log-return and realized variances.}

\item{LEVIER}{if `TRUE`, estime the MDSV model with leverage.}
}
\value{

}
\description{
Method for filtering the MDSV model on log-retruns and realized variances (uniquely or jointly).
}
\details{
The MDSVfilter help to simply filter a set of data with a predefined set of parameters. Like in \code{\link{MDSVfit}}, the 
likelihood calculation is performed in \code{C++} through the \pkg{Rcpp} package. 
According to Augustyniak et al., (2021), the para consist on a vector of :
\itemize{
    \item{\eqn{\omega}}{probability of success of the stationnary distribution of the Markov chain. (\eqn{0<\omega<1}).}
    \item{\eqn{a}}{highest persistance of the component of the MDSV process. (\eqn{0<a<1}).}
    \item{\eqn{b}}{decroissance rate of the persistances. (\eqn{1<b}).}
    \item{\eqn{\sigma}}{unconditionnal variance of the MDSV process.}
    \item{\eqn{\nu_0}}{states defined parameter. (\eqn{0<\nu_0<1}).}
    \item{\eqn{\gamma}}{parameter of the realized variances innovation. This parameter is required only if \eqn{ModelType=1} or \eqn{ModelType = 2}. (\eqn{0<\gamma}).}
    \item{\eqn{\xi}}{parameter of realized variances equation in joint estimation. This parameter is required only if \eqn{ModelType = 2}.}
    \item{\eqn{\varphi}}{parameter of realized variances equation in joint estimation. This parameter is required only if \eqn{ModelType = 2}.}
    \item{\eqn{\delta1}}{parameter of realized variances equation in joint estimation. This parameter is required only if \eqn{ModelType = 2}.}
    \item{\eqn{\delta2}}{parameter of realized variances equation in joint estimation. This parameter is required only if \eqn{ModelType = 2}.}
    \item{\eqn{l_1}}{Leverage effect parameter. This parameter is required only if \eqn{LEVIER = TRUE}. (\eqn{0<l_1}).}
    \item{\eqn{\theta_l}}{Leverage effect parameter. This parameter is required only if \eqn{LEVIER = TRUE}. (\eqn{0<\theta_l<1}).}
}
The leverage effect is taken into account according to the FHMV model (see Augustyniak et al., 2019). While filtering an
univariate realized variances data, log-returns are required to add leverage effect. 
AIC and BIC are computed using the formulas : 
\itemize{
\item{AIC : }{\eqn{L - k} }
\item{BIC : }{\eqn{L - (k/2) log(n)}}
}
where \eqn{L} is the log-likelihood, \eqn{k} is the number of parameters and \eqn{n} the number of observations in the dataset.
The \link[base]{class} of the output of this function is \code{MDSVfilter}. This class has a \link[base]{summary}, 
\link[base]{print} and \link[base]{plot} \link[utils]{methods} to summarize, print and plot the results. See 
\code{\link{summary.MDSVfilter}}, \code{\link{print.MDSVfilter}} and \code{\link{plot.MDSVfilter}} for more details.
}
\examples{
\dontrun{
# MDSV(N=2,K=3) without leverage on univariate log-returns S&P500
data     <- data(sp500)  # Data loading
N         <- 2           # Number of components
K         <- 3           # Number of states
ModelType <- 0           # Univariate log-returns
LEVIER    <- FALSE       # No leverage effect

# Model estimation
out_fit   <- MDSVfit(K=K,N=N,data=donne,ModelType = ModelType,LEVIER=LEVIER)
# Model filtering
para      <-out_fit$estimates # parameter
out_filter<- MDSVfilter(K = K, N = N, data = donne, para = para, ModelType = ModelType, LEVIER = LEVIER)
# Summary
summary(out_filter)
# Plot
plot(out)


# MDSV(N=3,K=3) with leverage on joint log-returns and realized variances S&P500
data     <- data(sp500)  # Data loading
N         <- 3           # Number of components
K         <- 3           # Number of states
ModelType <- 2           # Joint log-returns and realized variances
LEVIER    <- TRUE        # No leverage effect

para      <- c(omega = 0.52, a = 0.99, b = 2.77, sigma = 1.95, v0 = 0.72, 
              xi = -0.5, varphi = 0.93, delta1 = 0.93, delta2 = 0.04, shape = 2.10
              l = 0.78, theta = 0.876)
# Model estimation
out       <- MDSVfilter(K = K, N = N,data = donne, para = para, ModelType = ModelType, LEVIER = LEVIER)
# Summary
summary(out)
# Plot
plot(out)
}
}
\references{
Augustyniak, M., Bauwens, L., & Dufays, A. (2019). A new approach to volatility modeling: the factorial hidden Markov volatility model. 
\emph{Journal of Business & Economic Statistics}, 37(4), 696-709. \url{https://doi.org/10.1080/07350015.2017.1415910}

Augustyniak, M., Dufays, A., & Maoude, K.H.A. (2021). Multifractal Discrete Stochastic Volatility.
}
\seealso{
For fitting \code{\link{MDSVfit}}, bootstrap forecasting \code{\link{MDSVboot}} and rolling estimation and forecast \code{\link{MDSVroll}}.
}
