% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDSVfit.R
\name{MDSVfit}
\alias{MDSVfit}
\title{MDSV Fitting}
\usage{
MDSVfit(N, K, data, ModelType = 0, LEVIER = FALSE, start.pars = list(), ...)
}
\arguments{
\item{N}{An integer designing the number of components for the MDSV process}

\item{K}{An integer designing the number of states of each MDSV process component}

\item{data}{A univariate or bivariate data matrix. Can only be a matrix of 1 or 2 columns. If data has 2 columns, the first one has to be the log-returns and the second the realized variances.}

\item{ModelType}{An integer designing the type of model to be fit. \eqn{0} for univariate log-returns, \eqn{1} for univariate realized variances and \eqn{2} for joint log-return and realized variances.}

\item{LEVIER}{if \code{TRUE}, estime the MDSV model with leverage.}

\item{start.pars}{List of staring parameters for the optimization routine. These are not usually required unless the optimization has problems converging.}

\item{...}{Further options for the \code{\link{solnp}} solver of the \pkg{Rsolnp} package.}
}
\value{
A list consisting of:
\itemize{
    \item ModelType : type of model to be fitted.
    \item LEVIER : wheter the fit take the leverage effect into account or not.
    \item N : number of components for the MDSV process.
    \item K : number of states of each MDSV process component.
    \item convergence : 0 if convergence, 1 if not.
    \item estimates : estimated parameters.
    \item LogLikelihood : log-likelihood of the model on the data.
    \item AIC : Akaike Information Criteria of the model on the data.
    \item BIC : Bayesian Information Criteria of the model on the data.
    \item data : data use for the fitting.
}
}
\description{
Method for fitting the MDSV model on log-retruns and realized variances (uniquely or jointly).
}
\details{
The MDSV optimization routine set of feasible starting points which are used to initiate the MDSV recursion. The 
likelihood calculation is performed in \code{C++} through the \pkg{Rcpp} package. The optimization is perform using
the solnp solver of the \pkg{Rsolnp} package and additional options can be supply to the fonction. 
The leverage effect is taken into account according to the FHMV model (see Augustyniak et al., 2019). While fitting an
univariate realized variances data, log-returns are required to add leverage effect.
AIC and BIC are computed using the formulas : 
\itemize{
\item{AIC : }{\eqn{L - k}}
\item{BIC : }{\eqn{L - (k/2)*log(n)}}
}
where \eqn{L} is the log-likelihood, \eqn{k} is the number of parameters and \eqn{n} the number of observations in the dataset.
The \link[base]{class} of the output of this function is \code{\link{MDSVfit}}. This class has a \link[base]{summary}, 
\link[base]{print} and \link[base]{plot} \link[utils]{methods} to summarize, print and plot the results. See 
\code{\link{summary.MDSVfit}}, \code{\link{print.MDSVfit}} and \code{\link{plot.MDSVfit}} for more details.
}
\examples{
\dontrun{
# MDSV(N=2,K=3) without leverage on univariate log-returns S&P500
data(sp500)         # Data loading
N         <- 2      # Number of components
K         <- 3      # Number of states
ModelType <- 0      # Univariate log-returns
LEVIER    <- FALSE  # No leverage effect

# Model estimation
out       <- MDSVfit(K = K, N = N, data = sp500, ModelType = ModelType, LEVIER = LEVIER)
# Summary
summary(out)
# Plot
plot(out,c("dis","nic"))


# MDSV(N=3,K=3) with leverage on joint log-returns and realized variances NASDAQ
data(nasdaq)      # Data loading
N         <- 3    # Number of components
K         <- 3    # Number of states
ModelType <- 2    # Joint log-returns and realized variances
LEVIER    <- TRUE # No leverage effect

# Model estimation
out       <- MDSVfit(K = K, N = N, data = nasdaq, ModelType = ModelType, LEVIER = LEVIER)
# Summary
summary(out)
# Plot
plot(out,"nic")
}
}
\references{
Augustyniak, M., Bauwens, L., & Dufays, A. (2019). A new approach to volatility modeling: the factorial hidden Markov volatility model. 
\emph{Journal of Business & Economic Statistics}, 37(4), 696-709. \url{https://doi.org/10.1080/07350015.2017.1415910}
}
\seealso{
For filtering \code{\link{MDSVfilter}}, bootstrap forecasting \code{\link{MDSVboot}}, simulation \code{\link{MDSVsim}} and rolling estimation and forecast \code{\link{MDSVroll}}.
}
