% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDSVroll.R
\name{MDSVroll}
\alias{MDSVroll}
\title{MDSV Rolling estimates, volatility forecast and backtesting}
\usage{
MDSVroll(
  N,
  K,
  data,
  ModelType = 0,
  LEVIER = FALSE,
  n.ahead = 1,
  n.bootpred = 10000,
  forecast.length = 500,
  refit.every = 25,
  refit.window = "recursive",
  window.size = NULL,
  calculate.VaR = TRUE,
  VaR.alpha = c(0.01, 0.05),
  cluster = NULL,
  rseed = NA,
  ...
)
}
\arguments{
\item{N}{An integer designing the number of components for the MDSV process}

\item{K}{An integer designing the number of states of each MDSV process component}

\item{data}{A univariate or bivariate data matrix. Can only be a matrix of 1 or 2 columns. If data has 2 columns, the first one has to be the log-returns and the second the realized variances.}

\item{ModelType}{An integer designing the type of model to be fit. \eqn{0} for univariate log-returns, \eqn{1} for univariate realized variances and \eqn{2} for joint log-return and realized variances.}

\item{LEVIER}{if \code{TRUE}, estime the MDSV model with leverage.}

\item{n.ahead}{An integer designing the forecast horizon.}

\item{n.bootpred}{An integer designing the number of simulation based re-fits the model. Not relevant for one horizon forecast or for non-leverage type model.}

\item{rseed}{An integer use to initialize the random number generator for the resampling with replacement method (if not supplied take randomly).}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list consisting of:
\describe{
    \item{ModelType : }{type of model to be fitted.}
    \item{LEVIER : }{wheter the fit take the leverage effect into account or not.}
    \item{N : }{number of components for the MDSV process.}
    \item{K : }{number of states of each MDSV process component.}
    \item{estimates : }{estimated parameters.}
    \item{LogLikelihood : }{log-likelihood of the model on the data.}
    \item{AIC : }{Akaike Information Criteria of the model on the data.}
    \item{BIC : }{Bayesian Information Criteria of the model on the data.}
    \item{data : }{data use for the fitting.}
    \item{dates : }{vector or names of data designing the dates.}
    \item{n.ahead : }{integer designing the forecast horizon.}
    \item{n.bootpred : }{integer designing the number of simulation based re-fits used to generate the parameter distribution.}
    \item{rt_sim : }{matrix of log-returns forecast simulated where the row stand for the simulations and the columns for the horizon.}
    \item{rt2 : }{vector of mean by column of the square of rt_sim.}
    \item{rvt_sim : }{matrix of realized variances forecast simulated where the row stand for the simulations and the columns for the horizon.}
    \item{rvt : }{vector of mean by column of rvt_sim.}
}
}
\description{
Method for creating rolling estimates and volatility forecast from MDSV models with option for refitting every n periods 
with parallel functionality. The rolling estimate can be done using univariate log-returns or realized variances or using joint log-returns and realized variances.
}
\details{
The MDSVboot perform the forecasting of the model a different horizon. The forecasting is based on a close form where the estimation does not
involve leverage effect (see Hamilton, 1989. chapter 22 for hidden markov model forecasting). But to take into account the leverage effect,
the forecasting is perform by a bootstrap analysis. The innovations are bootstrapped using a standard normal distribution. This process 
is performed in \code{C++} through the \pkg{Rcpp} package. The leverage effect is taken into account according to the FHMV model 
(see Augustyniak et al., 2019). For the univariate realized variances forecasting, log-returns are required to add leverage effect. 

The \link[base]{class} of the output of this function is \code{MDSVboot}. This class has a \link[base]{summary} and 
\link[base]{print} \link[utils]{methods} to summarize and print the results. See 
\code{\link{summary.MDSVboot}}, \code{\link{print.MDSVboot}} and \code{\link{plot.MDSVboot}} for more details.
}
\examples{
\dontrun{
# MDSV(N=2,K=3) without leverage on univariate log-returns S&P500
data(sp500)         # Data loading
N         <- 2      # Number of components
K         <- 3      # Number of states
ModelType <- 0      # Univariate log-returns
LEVIER    <- FALSE  # No leverage effect

# Model estimation
out_fit   <- MDSVfit(K = K, N = N, data = sp500, ModelType = ModelType, LEVIER = LEVIER)
# Model forecasting (no bootstrapp is need as no leverage)
para      <-out_fit$estimates # parameter
out       <- MDSVboot(fit = out_fit, n.ahead = 100, rseed = 125)
# Summary
summary(out)


# MDSV(N=3,K=3) with leverage on joint log-returns and realized variances NASDAQ
data(nasdaq)       # Data loading
N         <- 3     # Number of components
K         <- 3     # Number of states
ModelType <- 2     # Joint log-returns and realized variances
LEVIER    <- TRUE  # No leverage effect

out_fit   <- MDSVfit(K = K, N = N, data = nasdaq, ModelType = ModelType, LEVIER = LEVIER)
# Model bootstrap forecasting
out       <- MDSVboot(fit = out_fit, n.ahead = 100, n.bootstrap = 10000, rseed = 349)
# Summary
summary(out)

}

}
\references{
Augustyniak, M., Bauwens, L., & Dufays, A. (2019). A new approach to volatility modeling: the factorial hidden Markov volatility model. 
\emph{Journal of Business & Economic Statistics}, 37(4), 696-709. \url{https://doi.org/10.1080/07350015.2017.1415910}
}
\seealso{
For fitting \code{\link{MDSVfit}}, filtering \code{\link{MDSVfilter}} and rolling estimation and forecast \code{\link{MDSVroll}}.
}
