% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omxc20-data.R
\docType{data}
\name{omxc20}
\alias{omxc20}
\title{OMX Copenhagen 20 Index}
\format{
Matrix of two columns where the first column is the log-returns and the second the realized variances.
}
\source{
\href{https://realized.oxford-man.ox.ac.uk/}{Oxford-Man Realized Library}
}
\usage{
data(omxc20)
}
\description{
Data from Oxford-Man institute of the OMX Copenhagen 20 Index from date "2005-10-03" to date "2021-02-24". 
Using close to close prices, log-retuns and realized variances are computed using formulas 
\eqn{r_t = log(P_{t})-log(P_{t-1})} and \eqn{rv_t = \sum_i r_{t,i}^2} where \eqn{r_{t,i}} stands 
for the intra-day log-returns of the date t and \eqn{P_t} stands for the price at day t.
}
\examples{
\dontrun{
data(omxc20)
}
}
\keyword{datasets}
