% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDSVroll.R
\name{plot.MDSVroll}
\alias{plot.MDSVroll}
\alias{print.plot.MDSVroll}
\title{Plot MDSV Rolling estimates, volatility forecast and backtesting}
\usage{
\method{plot}{MDSVroll}(x, plot.type = c("sigma", "VaR", "dens"), ...)

\method{print}{plot.MDSVroll}(x, ...)
}
\arguments{
\item{x}{An object of class \link{plot.MDSVroll}, output of the function \code{\link{plot.MDSVroll}}
or class \link{MDSVroll} of the function \code{\link{MDSVroll}}.}

\item{plot.type}{The type of plot to be draw.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
One or more of the following plots :
\itemize{
    \item sigma : Graph of the square of log-returns and/or realized variances with its 1-ahead forecasts.
    \item VaR : Graph showing the log-returns and the forecast conditional VaR at 1-ahead.
    \item dens : Graph of predicative density of log-retunrs or realized variances at 1-ahead.
}
}
\description{
Plot methods for the class \link{MDSVroll} as returned by the function \code{\link{MDSVroll}}.
}
\seealso{
For fitting \code{\link{MDSVfit}}, filtering \code{\link{MDSVfilter}}, bootstrap forecasting \code{\link{MDSVboot}} and rolling estimation and forecast \code{\link{MDSVroll}}.
}
