% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftse-data.R, R/nasdaq-data.R, R/sp500-data.R,
%   R/stoxx-data.R
\docType{data}
\name{sp500}
\alias{sp500}
\title{Arabidopsis QTL data on gravitropism}
\format{
Matrix of two columns where the first column is the log-returns and the second the realized variances.

Matrix of two columns where the first column is the log-returns and the second the realized variances.

Matrix of two columns where the first column is the log-returns and the second the realized variances.

Matrix of two columns where the first column is the log-returns and the second the realized variances.
}
\source{
\href{https://realized.oxford-man.ox.ac.uk/}{Oxford-Man Realized Library}

\href{https://realized.oxford-man.ox.ac.uk/}{Oxford-Man Realized Library}

\href{https://realized.oxford-man.ox.ac.uk/}{Oxford-Man Realized Library}

\href{https://realized.oxford-man.ox.ac.uk/}{Oxford-Man Realized Library}
}
\usage{
data(sp500)

data(sp500)

data(sp500)

data(sp500)
}
\description{
Data from a QTL experiment on gravitropism in
Arabidopsis, with data on 162 recombinant inbred lines (Ler x
Cvi). The outcome is the root tip angle (in degrees) at two-minute
increments over eight hours.

Data from a QTL experiment on gravitropism in
Arabidopsis, with data on 162 recombinant inbred lines (Ler x
Cvi). The outcome is the root tip angle (in degrees) at two-minute
increments over eight hours.

Data from a QTL experiment on gravitropism in
Arabidopsis, with data on 162 recombinant inbred lines (Ler x
Cvi). The outcome is the root tip angle (in degrees) at two-minute
increments over eight hours.

Data from a QTL experiment on gravitropism in
Arabidopsis, with data on 162 recombinant inbred lines (Ler x
Cvi). The outcome is the root tip angle (in degrees) at two-minute
increments over eight hours.
}
\keyword{datasets}
