% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDSVfit.R
\name{summary.MDSVfit}
\alias{summary.MDSVfit}
\alias{print.summary.MDSVfit}
\alias{print.MDSVfit}
\alias{plot.MDSVfit}
\alias{print.plot.MDSVfit}
\title{Summarize and Print MDSV Fitting}
\usage{
\method{summary}{MDSVfit}(object, ...)

\method{print}{summary.MDSVfit}(x, ...)

\method{print}{MDSVfit}(x, ...)

\method{plot}{MDSVfit}(x, plot.type = "both", ...)

\method{print}{plot.MDSVfit}(x, ...)
}
\arguments{
\item{object}{An object of class `MDSVfit`, output of the function \code{\link{MDSVfit}}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{An object of class `summary.MDSVfit`, output of the function \code{\link{summary.MDSVfit}}
or class `MDSVfit` of the function \code{\link{MDSVfit}}.}

\item{plot.type}{A character designing the type of plot. `"dis"` for the stationnary distribution of the volatilities,
`"nic"` for the New Impact Curve (see. Engle and Ng, 1993).}
}
\value{
A list consisting of:
    \item{ModelType}{type of model to be fitted.}
    \item{LEVIER}{wheter the fit take the leverage effect in account or not.}
    \item{N}{number of components for the MDSV process.}
    \item{K}{number of states of each MDSV process component.}
    \item{estimates}{estimated parameters.}
    \item{LogLikelihood}{log-likelihood of the model on the data.}
    \item{AIC}{Akaike Information Criteria of the model on the data.}
    \item{BIC}{Bayesian Information Criteria of the model on the data.}
    \item{data}{data use for the fitting.}
    \item{...}{further arguments passed to the function.}
}
\description{
Summary and print methods for the class `MDSVfit` as returned by the function \link{MDSVfit}.
}
