% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDSVroll.R
\name{summary.MDSVroll}
\alias{summary.MDSVroll}
\alias{print.summary.MDSVroll}
\alias{print.MDSVroll}
\title{Summarize and print MDSV Rolling estimates, volatility forecast and backtesting}
\usage{
\method{summary}{MDSVroll}(
  object,
  VaR.test = TRUE,
  Loss.horizon = c(1, 5, 10, 25, 50, 75, 100),
  Loss.window = 756,
  ...
)

\method{print}{summary.MDSVroll}(x, ...)

\method{print}{MDSVroll}(x, ...)
}
\arguments{
\item{object}{An object of class \link{MDSVroll}, output of the function \code{\link{MDSVroll}}.}

\item{VaR.test}{Whether to perform Value at Risk forecast backtesting.}

\item{Loss.horizon}{Horizon to summary the forecasts (cummulative and marginal).}

\item{Loss.window}{Window on which the forecasts are summarized.}

\item{...}{Further arguments passed to or from other methods.}

\item{x}{An object of class \link{summary.MDSVroll}, output of the function \code{\link{summary.MDSVroll}}
or class \link{MDSVroll} of the function \code{\link{MDSVroll}}.}
}
\value{
A list consisting of:
\itemize{
    \item N : number of components for the MDSV process.
    \item K : number of states of each MDSV process component.
    \item ModelType : type of models fitted.
    \item LEVIER : wheter the fit take the leverage effect into account or not.
    \item n.ahead : integer designing the forecast horizon.
    \item forecast.length : length of the total forecast for which out of sample data from the dataset will be used for testing.
    \item refit.every : Determines every how many periods the model is re-estimated.
    \item refit.window : Whether the refit is done on an expanding window including all the previous data or a moving window where all previous 
data is used for the first estimation and then moved by a length equal to refit.every (unless the window.size option is used instead).
    \item window.size : If not NULL, determines the size of the moving window in the rolling estimation, which also determines the first point used.
    \item calculate.VaR : Whether to calculate forecast Value at Risk during the estimation.
    \item VaR.alpha : The Value at Risk tail level to calculate.
    \item cluster : A cluster object created by calling makeCluster from the parallel package.
    \item data : data use for the fitting.
    \item dates : vector or names of data designing the dates.
    \item estimates : matrix of all the parameters estimates at each date.
    \item prevision : matrix of all prevision made a each date.
    \item VaR.test : Whether to perform Value at Risk forecast backtesting.
    \item Loss.horizon : Horizon to summary the forecasts.
    \item Loss.window : Window on which the forecasts are summarized.
    \item Loss : Matrice containing the forecasts summary.
}
}
\description{
Summary and print methods for the class \link{MDSVroll} as returned by the function \code{\link{MDSVroll}}.
}
\details{
The \code{\link{summary.MDSVroll}} function compute the Root Mean Square Error, the Mean Average Error and the Quasi-Likehood 
error to summarize the forecasts. Those loss functions are compute for cummulative (by horizon) and marginal forecasts. 
For univariate realized variances model and joint log-returns and realized variances model, the loss functions are computed for
the realized variances and for the univariate log-returns model, the loss functions are computed for the log-returns.
For the Value-at-Risk basktest, the unconditionnal coverage test (see. Kupiec), the independance test (see Christoffersen) and the 
conditional coverage test (see Christoffersen and ) are performed.
}
\seealso{
For fitting \code{\link{MDSVfit}}, filtering \code{\link{MDSVfilter}}, bootstrap forecasting \code{\link{MDSVboot}} and rolling estimation and forecast \code{\link{MDSVroll}}.
}
