% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDSVfit.R
\name{summary.MDSVfit}
\alias{summary.MDSVfit}
\alias{print.summary.MDSVfit}
\alias{print.MDSVfit}
\alias{plot.MDSVfit}
\alias{print.plot.MDSVfit}
\title{Summarize, print and plot MDSV Fitting}
\usage{
\method{summary}{MDSVfit}(object, ...)

\method{print}{summary.MDSVfit}(x, ...)

\method{print}{MDSVfit}(x, ...)

\method{plot}{MDSVfit}(x, plot.type = c("dis", "nic"), ...)

\method{print}{plot.MDSVfit}(x, ...)
}
\arguments{
\item{object}{An object of class \link{MDSVfit}, output of the function \code{\link{MDSVfit}}.}

\item{...}{Further arguments passed to or from other methods.}

\item{x}{An object of class \link{summary.MDSVfit}, output of the function \code{\link{summary.MDSVfit}},
class \link{MDSVfit} of the function \code{\link{MDSVfit}} or class \link{plot.MDSVfit} of the function \code{\link{plot.MDSVfit}}.}

\item{plot.type}{A character designing the type of plot. \code{dis} for the stationnary distribution of the volatilities,
\code{nic} for the New Impact Curve (see. Engle and Ng, 1993).}
}
\value{
A list consisting of:
\itemize{
    \item ModelType : type of model to be fitted.
    \item LEVIER : wheter the fit take the leverage effect into account or not.
    \item N : number of components for the MDSV process.
    \item K : number of states of each MDSV process component.
    \item convergence : 0 if convergence, 1 if not.
    \item estimates : estimated parameters.
    \item LogLikelihood : log-likelihood of the model on the data.
    \item AIC : Akaike Information Criteria of the model on the data.
    \item BIC : Bayesian Information Criteria of the model on the data.
    \item data : data use for the fitting.
    \item ... : further arguments passed to the function.
}
}
\description{
Summary, print and plot methods for the class \link{MDSVfit} as returned by the function \code{\link{MDSVfit}}.
}
\details{
\code{dis} as argument \code{plot.type} lead to plot the stationnary distribution of the Markov chain process MDSV. The leverage
effect is not took into account for that plot.
}
\references{
Engle, R. F., & Ng, V. K. (1993). Measuring and testing the impact of news on volatility. 
\emph{The journal of finance}, 48(5), 1749-1778. \url{ https://doi.org/10.1111/j.1540-6261.1993.tb05127.x}
}
\seealso{
For fitting \code{\link{MDSVfit}}, filtering \code{\link{MDSVfilter}}, bootstrap forecasting \code{\link{MDSVboot}} and rolling estimation and forecast \code{\link{MDSVroll}}.
}
